+++ src/test/ide.test.ts
@@ -4,10 +4,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import {inspect} from 'util';
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
+import {drawSquiggle, OffsetToPositionConverter} from '../error.js';
 import {IdeAnalyzer} from '../ide.js';
 import {WireitTestRig} from './util/test-rig.js';
+import * as url from 'url';
+import {removeAciiColors} from './util/colors.js';
 
 const test = suite<{rig: WireitTestRig}>();
 
@@ -147,6 +151,345 @@ test('we can get cyclic dependency errors', async ({rig}) => {
   );
   await assertDiagnostics(ide, {
     [rig.resolve('package.json')]: [`Cycle detected in dependencies of "a".`],
+  });
+});
+
+async function assertDefinition(
+  ide: IdeAnalyzer,
+  options: {
+    path: string;
+    contentsWithPipe: string;
+    expected:
+      | undefined
+      | {
+          target: string;
+          targetSelection: string;
+          originSelection: string;
+        };
+  }
+) {
+  const offset = options.contentsWithPipe.indexOf('|');
+  const contents =
+    options.contentsWithPipe.slice(0, offset) +
+    options.contentsWithPipe.slice(offset + 1);
+  ide.setOpenFileContents(options.path, contents);
+  const sourceFile = await ide.getPackageJsonForTest(options.path);
+  if (sourceFile == null) {
+    throw new Error(`could not get source file`);
+  }
+  const sourceConverter = OffsetToPositionConverter.get(sourceFile.jsonFile);
+  const definitions = await ide.getDefinition(
+    options.path,
+    sourceConverter.toIdePosition(offset)
+  );
+  if (definitions === undefined) {
+    if (options.expected === undefined) {
+      return; // No definition, as expected.
+    }
+    throw new Error(
+      `Expected to find a definition matching \n${inspect(options.expected)}`
+    );
+  }
+  // currently we always expect one definition
+  assert.equal(definitions.length, 1);
+  const definition = definitions[0];
+  if (options.expected === undefined) {
+    throw new Error(
+      `Expected no definition, but got one: ${inspect(definition)}`
+    );
+  }
+  const targetFile = await ide.getPackageJsonForTest(
+    url.fileURLToPath(definition.targetUri)
+  );
+  if (targetFile == null) {
+    throw new Error(`Could not load target file`);
+  }
+  const targetConverter = OffsetToPositionConverter.get(targetFile.jsonFile);
+  const targetSquiggle = drawSquiggle(
+    {
+      file: targetFile.jsonFile,
+      range: targetConverter.ideRangeToRange(definition.targetRange),
+    },
+    0
+  );
+  assertSquiggleEquals(targetSquiggle, options.expected.target);
+
+  const targetSelectionSquiggle = drawSquiggle(
+    {
+      file: targetFile.jsonFile,
+      range: targetConverter.ideRangeToRange(definition.targetSelectionRange),
+    },
+    0
+  );
+  assertSquiggleEquals(
+    targetSelectionSquiggle,
+    options.expected.targetSelection
+  );
+  if (definition.originSelectionRange == null) {
+    throw new Error(`No iriginSelectionRange returned`);
+  }
+  const sourceSelectionSquiggle = drawSquiggle(
+    {
+      file: sourceFile.jsonFile,
+      range: sourceConverter.ideRangeToRange(definition.originSelectionRange),
+    },
+    2
+  );
+  assertSquiggleEquals(
+    sourceSelectionSquiggle,
+    options.expected.originSelection
+  );
+}
+
+function assertSquiggleEquals(actual: string, expected: string) {
+  actual = removeAciiColors(actual);
+  if (actual.trim() !== expected.trim()) {
+    console.log(`Copy pastable output:\n${actual}`);
+  }
+  assert.equal(actual.trim(), expected.trim());
+}
+
+test('we can get the definition for a same file dependency', async ({rig}) => {
+  const ide = new IdeAnalyzer();
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+          b: 'wireit',
+        },
+        wireit: {
+          a: {
+            dependencies: ['|b'],
+          },
+          b: {
+            command: 'echo',
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: {
+      target: `
+    "b": {
+    ~~~~~~
+      "command": "echo"
+~~~~~~~~~~~~~~~~~~~~~~~
+    }
+~~~~~`,
+      targetSelection: `
+    "b": {
+    ~~~`,
+      originSelection: `
+          "b"
+          ~~~`,
+    },
+  });
+});
+
+test(`we jump to the scripts section for a vanilla script`, async ({rig}) => {
+  const ide = new IdeAnalyzer();
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+          b: 'echo',
+        },
+        wireit: {
+          a: {
+            dependencies: ['|b'],
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: {
+      target: `
+    "b": "echo"
+    ~~~~~~~~~~~`,
+      targetSelection: `
+    "b": "echo"
+    ~~~`,
+      originSelection: `
+          "b"
+          ~~~`,
+    },
+  });
+});
+
+test(`we don't get definitions for non-dep locations`, async ({rig}) => {
+  const ide = new IdeAnalyzer();
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+          b: 'wireit',
+        },
+        wireit: {
+          a: {
+            dependencies: ['b'],
+            foo: 'b|ar',
+          },
+          b: {
+            command: 'echo',
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: undefined,
+  });
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+          b: 'wireit',
+        },
+        wireit: {
+          a: {
+            dependencies: ['b'],
+            foo: 'b|ar',
+          },
+          b: {
+            command: 'echo',
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: undefined,
+  });
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+          b: 'wireit',
+        },
+        wireit: {
+          a: {
+            ['depen|dencies']: ['b'],
+          },
+          b: {
+            command: 'echo',
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: undefined,
+  });
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+          b: 'wireit',
+        },
+        wireit: {
+          ['|a']: {
+            dependencies: ['b'],
+          },
+          b: {
+            command: 'echo',
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: undefined,
+  });
+});
+
+test(`we can jump to definitions across files`, async ({rig}) => {
+  const ide = new IdeAnalyzer();
+  await rig.write('child/package.json', {scripts: {b: 'echo'}});
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wireit',
+        },
+        wireit: {
+          a: {
+            dependencies: ['./chil|d:b'],
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: {
+      target: `
+    "b": "echo"
+    ~~~~~~~~~~~`,
+      targetSelection: `
+    "b": "echo"
+    ~~~`,
+      originSelection: `
+          "./child:b"
+          ~~~~~~~~~~~`,
+    },
+  });
+});
+
+test('can jump from scripts section to wireit config', async ({rig}) => {
+  const ide = new IdeAnalyzer();
+  await assertDefinition(ide, {
+    path: rig.resolve('package.json'),
+    contentsWithPipe: JSON.stringify(
+      {
+        scripts: {
+          a: 'wir|eit',
+          b: 'wireit',
+        },
+        wireit: {
+          a: {
+            dependencies: ['b'],
+          },
+          b: {
+            command: 'echo',
+          },
+        },
+      },
+      null,
+      2
+    ),
+    expected: {
+      target: `
+    "a": {
+    ~~~~~~
+      "dependencies": [
+~~~~~~~~~~~~~~~~~~~~~~~
+        "b"
+~~~~~~~~~~~
+      ]
+~~~~~~~
+    },
+~~~~~`,
+      targetSelection: `
+    "a": {
+    ~~~`,
+      originSelection: `
+      "a": "wireit",
+      ~~~~~~~~~~~~~`,
+    },
   });
 });